+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 - Add strength and angle sensor for Xiaomi DJT11LM
+- Add smoke density, self test and sensivity support for Xiaomi JTYJ-GD-01LM/BW
 
 ## 0.2.0
 - Support for "stop" command for blinds adapter